Allow float input for square calculation
Changed input type from int to float for square calculation.

+++ 2.1.6.py
@@ -1,6 +1,6 @@
 # to find the square of the number entered by the user
 print ("WELCOME TO THE SQUARE CALCULATOR")
-a = int(input("enter the number you want the square of: "))
+a = float(input("enter the number you want the square of: "))
 print ("Therefore the square is", a*a)
 ''' ANOTHER WAY IS 
 print ("Therefore the square is", a**2)'''
